+++ apps/site/models/userprofile.py
@@ -2,7 +2,7 @@ from django.db.models import signals
 from django.contrib.gis.db import models
 from django.contrib.auth.models import User
 from localground.apps.site.models.permissions import ObjectAuthority
-from localground.apps.site.models.groups import Project
+from localground.apps.site.models.project import Project
 from datetime import datetime
 from django.conf import settings
